Update README.md
This is a project of computer graphics, which is done by using C++ and openGL. Ii's not a fully complete project. Because the collision part is not working. The problem is, in this game the position of the ball is not fix, it's moving both X axis and Y axis. thats why it's very difficult to get the exact value of the ball. Which i need to develop the collision part. Otherwise the game is a coplete game.